correcting value

+++ include/ADC_private.h
@@ -34,6 +34,6 @@
 #define ADCH              *((volatile u8*) 0x25 )
 #define ADCL              *((volatile u8*) 0x24 )
 #define SFIOR             *((volatile u8*) 0x50 )
-#define ADC_VAL           *((volatile u8*) 0x24 )  
+#define ADC_VAL           *((volatile u16*) 0x24 )  
 
 #endif
